finally got the app to work in the right components, naming could be better
