+++ src/api/Data.js
@@ -1,0 +1,100 @@
+export const antwortenF1 = [
+    {
+      WertRot: 0,
+      WertGruen: 0
+    },
+    {
+      Wert1: "+",
+      Wert2: "-",
+      Wert3: "0",
+      WertRot: -1,
+      WertGruen: -1
+    },
+    {
+      Wert1: "+",
+      Wert2: "0",
+      Wert3: "-",
+      WertRot: 0,
+      WertGruen: -1
+    },
+    {
+      Wert1: "0",
+      Wert2: "+",
+      Wert3: "-",
+      WertRot: 1,
+      WertGruen: 0
+    },
+    {
+      Wert1: "-",
+      Wert2: "+",
+      Wert3: "0",
+      WertRot: 1,
+      WertGruen: 1
+    },
+    {
+      Wert1: "0",
+      Wert2: "-",
+      Wert3: "+",
+      WertRot: -1,
+      WertGruen: 0
+    },
+    {
+      Wert1: "-",
+      Wert2: "0",
+      Wert3: "+",
+      WertRot: 0,
+      WertGruen: 1
+    }
+  ];
+   
+  export const antwortenF2 = [
+    {
+      Wert1: "+",
+      Wert2: "-",
+      Wert3: "0",
+      WertRot: 0,
+      WertGruen: -1
+    },
+    {
+      Wert1: "+",
+      Wert2: "0",
+      Wert3: "-",
+      WertRot: 0,
+      WertGruen: 1
+    },
+    {
+      Wert1: "0",
+      Wert2: "+",
+      Wert3: "-",
+      WertRot: 1,
+      WertGruen: 1
+    },
+    {
+      Wert1: "-",
+      Wert2: "+",
+      Wert3: "0",
+      WertRot: -1,
+      WertGruen: -1
+    },
+    {
+      Wert1: "0",
+      Wert2: "-",
+      Wert3: "+",
+      WertRot: 1,
+      WertGruen: 0
+    },
+    {
+      Wert1: "-",
+      Wert2: "0",
+      Wert3: "+",
+      WertRot: -1,
+      WertGruen: 1
+    }
+  ];
+   
+  export const fragen = ["Wie sehen sie sich?", "Wie sehen sie sich nicht?"];
+   
+  export const antwortenFrage1 = ["dick", "dünn", "hässlich"];
+   
+  export const antwortenFrage2 = ["dickpimmelig", "schlau", "reich"];
+
